+++ test/cli/test_serve.py
@@ -13,13 +13,13 @@ class ServerCliTest(CliDataTest):
     def test_config(self):
         result = self.invoke_cli(["serve", "--config", "x.yml", "pippo.zarr"])
         self.assertEqual('Error: CONFIG and CUBES cannot be used at the same time.\n',
-                         result.output[-58:])
+                         result.stderr[-58:])
         self.assertEqual(1, result.exit_code)
 
         with ChangeEnv({CONFIG_ENV_VAR: "x.yml"}):
             result = self.invoke_cli(["serve"])
             self.assertEqual('Error: Configuration file not found: x.yml\n',
-                             result.output[-44:])
+                             result.stderr[-44:])
             self.assertEqual(1, result.exit_code)
 
     def test_base_dir(self):
@@ -29,13 +29,13 @@ class ServerCliTest(CliDataTest):
         with ChangeEnv({BASE_ENV_VAR: None}):
             result = self.invoke_cli(["serve", "--config", config_path, "--base-dir", "pippo/gnarz"])
             self.assertEqual('Error: Base directory not found: pippo/gnarz\n',
-                             result.output[-45:])
+                             result.stderr[-45:])
             self.assertEqual(1, result.exit_code)
 
         with ChangeEnv({BASE_ENV_VAR: 'pippo/configs'}):
             result = self.invoke_cli(["serve", "--config", config_path])
             self.assertEqual('Error: Base directory not found: pippo/configs\n',
-                             result.output[-47:])
+                             result.stderr[-47:])
             self.assertEqual(1, result.exit_code)
 
     def test_show(self):
@@ -47,7 +47,7 @@ class ServerCliTest(CliDataTest):
             self.assertEqual(f'Error: Option "--show": In order to run the viewer, '
                              f'set environment variable {VIEWER_ENV_VAR} to a valid '
                              f'xcube-viewer deployment or build directory.\n',
-                             result.output[-150:])
+                             result.stderr[-150:])
             self.assertEqual(2, result.exit_code)
 
         with ChangeEnv({VIEWER_ENV_VAR: "pippo/viewer"}):
@@ -55,7 +55,7 @@ class ServerCliTest(CliDataTest):
             self.assertEqual(f'Error: Option "--show": Viewer path set by environment '
                              f'variable {VIEWER_ENV_VAR} '
                              f'must be a directory: pippo/viewer\n',
-                             result.output[-116:])
+                             result.stderr[-116:])
             self.assertEqual(2, result.exit_code)
 
     def test_help(self):